requirements.txt
# Copyright (C) 2017 Corey Farmer
#
# This file is part of Tesseract OCR.
#
# Hermes is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# Hermes is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with Hermes.  If not, see <http://www.gnu.org/licenses/>.

# This is an implicit value, here for clarity
# This file is used by python pip to install the following dependencies

git+https://github.com/jflesch/pyocr.git/releases/0.4.6
wand==0.4.6
docopt==0.6.2
atxpdf==0.1.3
python-daemon==2.1.2
